Make setup.py more robust.
According to setuptools documentation [1], "include_package_data=True"
bit only supports CVS or SVN out of the box for setuptools.  Since the
project is hosted on github, this breaks installation for people that
install from the git repository via pip or any git checkout of the
project.

The MANIFEST.in file appears to be necessary to generate the proper
tarball using "python setup.py sdist".

In order to support all use cases, we have to manually specify the data
directories in setup.py.

[1] http://peak.telecommunity.com/DevCenter/setuptools#including-data-files