Fixing Cloudbase Cluster Launch Issues
Added enhanced testing to ensure cloudbase is up and running once the
start-cloudbase code has run. This fixes launching clusters with
PROVIDER=cloudbase and "start-cloudbase" calls. Retry on cloudbase
setup is called once a minute if a failure to start occurs. The root
cause of the issue seems to be slow AMI tablet servers not accepting
the start-server command from the master node on the first try (even on
slow AMIs full start seems achieved in under 5 tries now).